[bug] copy returns a closed handle
copy call returns a closed file handler, instead we can query for the
blob after insertion and use the returned file handle to be consumed for
upload.
A better solution would be to "pipe" the writes into the database into
the upload, but that involves solving a larger set of issues to be done
later.